+++ lib/hammer.ex
@@ -2,8 +2,8 @@ defmodule Hammer do
   @moduledoc """
   Documentation for Hammer module.
 
-  This is the main API for the Hammer rate-limiter. This module assumes an appropriate
-  backend has been configured.
+  This is the main API for the Hammer rate-limiter. This module assumes a
+  backend pool has been started, most likely by the Hammer application.
   """
 
   alias Hammer.Utils
@@ -16,8 +16,9 @@ defmodule Hammer do
   Check if the action you wish to perform is within the bounds of the rate-limit.
 
   Args:
-  - `id`: String name of the bucket. Usually the bucket name is comprised of some fixed prefix,
-  with some dynamic string appended, such as an IP address or user id.
+  - `id`: String name of the bucket. Usually the bucket name is comprised of
+  some fixed prefix, with some dynamic string appended, such as an IP address or
+  user id.
   - `scale_ms`: Integer indicating size of bucket in milliseconds
   - `limit`: Integer maximum count of actions within the bucket
 
@@ -60,20 +61,69 @@ defmodule Hammer do
     end
   end
 
+  @spec check_rate_inc(
+          id :: String.t(),
+          scale_ms :: integer,
+          limit :: integer,
+          increment :: integer
+        ) ::
+          {:allow, count :: integer}
+          | {:deny, limit :: integer}
+          | {:error, reason :: any}
+  @doc """
+  Same as check_rate/3, but allows the increment number to be specified.
+  This is useful for limiting apis which have some idea of 'cost', where the cost
+  of each hit can be specified.
+  """
+  def check_rate_inc(id, scale_ms, limit, increment) do
+    check_rate_inc(:single, id, scale_ms, limit, increment)
+  end
+
+  @spec check_rate_inc(
+          backend :: atom,
+          id :: String.t(),
+          scale_ms :: integer,
+          limit :: integer,
+          increment :: integer
+        ) ::
+          {:allow, count :: integer}
+          | {:deny, limit :: integer}
+          | {:error, reason :: any}
+  @doc """
+  Same as check_rate_inc/4, but allows specifying a backend.
+  """
+  def check_rate_inc(backend, id, scale_ms, limit, increment) do
+    {stamp, key} = Utils.stamp_key(id, scale_ms)
+
+    case call_backend(backend, :count_hit, [key, stamp, increment]) do
+      {:ok, count} ->
+        if count > limit do
+          {:deny, limit}
+        else
+          {:allow, count}
+        end
+
+      {:error, reason} ->
+        {:error, reason}
+    end
+  end
+
   @spec inspect_bucket(id :: String.t(), scale_ms :: integer, limit :: integer) ::
           {:ok,
            {count :: integer, count_remaining :: integer, ms_to_next_bucket :: integer,
             created_at :: integer | nil, updated_at :: integer | nil}}
           | {:error, reason :: any}
   @doc """
-  Inspect bucket to get count, count_remaining, ms_to_next_bucket, created_at, updated_at.
-  This function is free of side-effects and should be called with the same arguments you
-  would use for `check_rate` if you intended to increment and check the bucket counter.
+  Inspect bucket to get count, count_remaining, ms_to_next_bucket, created_at,
+  updated_at. This function is free of side-effects and should be called with
+  the same arguments you would use for `check_rate` if you intended to increment
+  and check the bucket counter.
 
   Arguments:
 
-  - `id`: String name of the bucket. Usually the bucket name is comprised of some fixed prefix,
-  with some dynamic string appended, such as an IP address or user id.
+  - `id`: String name of the bucket. Usually the bucket name is comprised of
+    some fixed prefix,with some dynamic string appended, such as an IP address
+    or user id.
   - `scale_ms`: Integer indicating size of bucket in milliseconds
   - `limit`: Integer maximum count of actions within the bucket
 
@@ -164,8 +214,8 @@ defmodule Hammer do
   - `scale_ms`: Integer indicating size of bucket in milliseconds
   - `limit`: Integer maximum count of actions within the bucket
 
-  Returns a function which accepts an `id` suffix, which will be combined with the `id_prefix`.
-  Calling this returned function is equivalent to:
+  Returns a function which accepts an `id` suffix, which will be combined with
+  the `id_prefix`. Calling this returned function is equivalent to:
   `Hammer.check_rate("\#{id_prefix}\#{id}", scale_ms, limit)`
 
   Example:
@@ -202,7 +252,14 @@ defmodule Hammer do
   end
 
   defp call_backend(which, function, args) do
+    pool = Utils.pool_name(which)
     backend = Utils.get_backend_module(which)
-    apply(backend, function, args)
+
+    :poolboy.transaction(
+      pool,
+      fn pid -> apply(backend, function, [pid | args]) end,
+      # TODO: make timeout configurable
+      60000
+    )
   end
 end